Update Report.md
Updated the links to go to the LucidChart site, instead of the images we have in the repository. I started to look into the "Security Related Configuration" section, but it was late and it may not be what was being asked.  I entered a description for the data leakage section too.  Thanks.